+++ dummyserver/testcase.py
@@ -301,7 +301,6 @@ class ConnectionMarker:
         """
 
         orig_request = HTTPConnection.request
-        orig_request_chunked = HTTPConnection.request_chunked
 
         def call_and_mark(target: Callable[..., None]) -> Callable[..., None]:
             def part(self: HTTPConnection, *args: Any, **kwargs: Any) -> None:
@@ -312,9 +311,6 @@ class ConnectionMarker:
 
         with monkeypatch.context() as m:
             m.setattr(HTTPConnection, "request", call_and_mark(orig_request))
-            m.setattr(
-                HTTPConnection, "request_chunked", call_and_mark(orig_request_chunked)
-            )
             yield
 
     @classmethod