+++ noxfile.py
@@ -16,7 +16,7 @@ SOURCE_FILES = [
 
 def tests_impl(
     session: nox.Session,
-    extras: str = "socks,secure,brotli",
+    extras: str = "socks,secure,brotli,zstd",
     byte_string_comparisons: bool = True,
 ) -> None:
     # Install deps and the package itself.
@@ -92,6 +92,12 @@ def downstream_botocore(session: nox.Session) -> None:
     session.cd(tmp_dir)
     git_clone(session, "https://github.com/boto/botocore")
     session.chdir("botocore")
+    session.run(
+        "git",
+        "apply",
+        f"{root}/ci/0001-Mark-100-Continue-tests-as-failing.patch",
+        external=True,
+    )
     session.run("git", "rev-parse", "HEAD", external=True)
     session.run("python", "scripts/ci/install")
 
@@ -127,19 +133,6 @@ def downstream_requests(session: nox.Session) -> None:
 @nox.session()
 def format(session: nox.Session) -> None:
     """Run code formatters."""
-    session.install("pre-commit")
-    session.run("pre-commit", "--version")
-
-    process = subprocess.run(
-        ["pre-commit", "run", "--all-files"],
-        env=session.env,
-        text=True,
-        stdout=subprocess.PIPE,
-        stderr=subprocess.STDOUT,
-    )
-    # Ensure that pre-commit itself ran successfully
-    assert process.returncode in (0, 1)
-
     lint(session)
 
 
@@ -168,7 +161,7 @@ def mypy(session: nox.Session) -> None:
 @nox.session
 def docs(session: nox.Session) -> None:
     session.install("-r", "docs/requirements.txt")
-    session.install(".[socks,secure,brotli]")
+    session.install(".[socks,secure,brotli,zstd]")
 
     session.chdir("docs")
     if os.path.exists("_build"):