+++ src/urllib3/connection.py
@@ -4,7 +4,6 @@ import os
 import re
 import socket
 import warnings
-from copy import copy
 from http.client import HTTPConnection as _HTTPConnection
 from http.client import HTTPException as HTTPException  # noqa: F401
 from socket import timeout as SocketTimeout
@@ -30,7 +29,7 @@ if TYPE_CHECKING:
     from .util.ssltransport import SSLTransport
 
 from .util.timeout import _DEFAULT_TIMEOUT, _TYPE_TIMEOUT, Timeout
-from .util.util import to_bytes, to_str
+from .util.util import to_str
 
 try:  # Compiled with SSL?
     import ssl
@@ -52,6 +51,7 @@ from .exceptions import (
     SystemTimeWarning,
 )
 from .util import SKIP_HEADER, SKIPPABLE_HEADERS, connection, ssl_
+from .util.request import body_to_chunks
 from .util.ssl_ import assert_fingerprint as _assert_fingerprint
 from .util.ssl_ import (
     create_urllib3_context,
@@ -74,7 +74,7 @@ port_by_scheme = {"http": 80, "https": 443}
 
 # When it comes time to update this value as a part of regular maintenance
 # (ie test_recent_date is failing) update it to ~6 months before the current date.
-RECENT_DATE = datetime.date(2020, 7, 1)
+RECENT_DATE = datetime.date(2022, 1, 1)
 
 _CONTAINS_CONTROL_CHAR_RE = re.compile(r"[^-!#$%&'*+.^_`|~0-9a-zA-Z]")
 
@@ -85,6 +85,8 @@ _TYPE_BODY = Union[bytes, IO[Any], Iterable[bytes], str, "MultipartEncoder"]
 class ProxyConfig(NamedTuple):
     ssl_context: Optional["ssl.SSLContext"]
     use_forwarding_for_https: bool
+    assert_hostname: Union[None, str, "Literal[False]"]
+    assert_fingerprint: Optional[str]
 
 
 class HTTPConnection(_HTTPConnection):
@@ -127,6 +129,7 @@ class HTTPConnection(_HTTPConnection):
     #: certificate. If no HTTPS proxy is being used will be ``None``.
     proxy_is_verified: Optional[bool] = None
 
+    blocksize: int
     source_address: Optional[Tuple[str, int]]
     socket_options: Optional[connection._TYPE_SOCKET_OPTIONS]
     _tunnel_host: Optional[str]
@@ -289,25 +292,72 @@ class HTTPConnection(_HTTPConnection):
         url: str,
         body: Optional[_TYPE_BODY] = None,
         headers: Optional[Mapping[str, str]] = None,
+        chunked: bool = False,
     ) -> None:
+
         if headers is None:
             headers = {}
-        else:
-            # Avoid modifying the headers passed into .request()
-            headers = copy(headers)
-            # Don't send bytes keys to httplib to avoid bytes/str comparison
-            # HTTPHeaderDict is already safe, but other types are not
-            for key, value in list(headers.items()):
-                if isinstance(key, bytes):
-                    headers.pop(key)
-                    # httplib would have decoded to latin-1 anyway
-                    headers[key.decode("latin-1")] = value
+        header_keys = frozenset(to_str(k.lower()) for k in headers)
+        skip_accept_encoding = "accept-encoding" in header_keys
+        skip_host = "host" in header_keys
+        self.putrequest(
+            method, url, skip_accept_encoding=skip_accept_encoding, skip_host=skip_host
+        )
 
-        if "user-agent" not in (to_str(k.lower()) for k in headers):
-            updated_headers = {"User-Agent": _get_default_user_agent()}
-            updated_headers.update(headers)
-            headers = updated_headers
-        super().request(method, url, body=body, headers=headers)
+        # Transform the body into an iterable of sendall()-able chunks
+        # and detect if an explicit Content-Length is doable.
+        chunks_and_cl = body_to_chunks(body, method=method, blocksize=self.blocksize)
+        chunks = chunks_and_cl.chunks
+        content_length = chunks_and_cl.content_length
+
+        # When chunked is explicit set to 'True' we respect that.
+        if chunked:
+            if "transfer-encoding" not in header_keys:
+                self.putheader("Transfer-Encoding", "chunked")
+        else:
+            # Detect whether a framing mechanism is already in use. If so
+            # we respect that value, otherwise we pick chunked vs content-length
+            # depending on the type of 'body'.
+            if "content-length" in header_keys:
+                chunked = False
+            elif "transfer-encoding" in header_keys:
+                chunked = True
+
+            # Otherwise we go off the recommendation of 'body_to_chunks()'.
+            else:
+                chunked = False
+                if content_length is None:
+                    if chunks is not None:
+                        chunked = True
+                        self.putheader("Transfer-Encoding", "chunked")
+                else:
+                    self.putheader("Content-Length", str(content_length))
+
+        # Now that framing headers are out of the way we send all the other headers.
+        if "user-agent" not in header_keys:
+            self.putheader("User-Agent", _get_default_user_agent())
+        for header, value in headers.items():
+            self.putheader(header, value)
+        self.endheaders()
+
+        # If we're given a body we start sending that in chunks.
+        if chunks is not None:
+            for chunk in chunks:
+                # Sending empty chunks isn't allowed for TE: chunked
+                # as it indicates the end of the body.
+                if not chunk:
+                    continue
+                if isinstance(chunk, str):
+                    chunk = chunk.encode("utf-8")
+                if chunked:
+                    self.send(b"%x\r\n%b\r\n" % (len(chunk), chunk))
+                else:
+                    self.send(chunk)
+
+        # Regardless of whether we have a body or not, if we're in
+        # chunked mode we want to send an explicit empty chunk.
+        if chunked:
+            self.send(b"0\r\n\r\n")
 
     def request_chunked(
         self,
@@ -320,39 +370,7 @@ class HTTPConnection(_HTTPConnection):
         Alternative to the common request method, which sends the
         body with chunked encoding and not as one block
         """
-        if headers is None:
-            headers = {}
-        header_keys = {to_str(k.lower()) for k in headers}
-        skip_accept_encoding = "accept-encoding" in header_keys
-        skip_host = "host" in header_keys
-        self.putrequest(
-            method, url, skip_accept_encoding=skip_accept_encoding, skip_host=skip_host
-        )
-        if "user-agent" not in header_keys:
-            self.putheader("User-Agent", _get_default_user_agent())
-        for header, value in headers.items():
-            self.putheader(header, value)
-        if "transfer-encoding" not in header_keys:
-            self.putheader("Transfer-Encoding", "chunked")
-        self.endheaders()
-
-        if body is not None:
-            if isinstance(body, (str, bytes)):
-                body = (to_bytes(body),)
-            for chunk in body:
-                if not chunk:
-                    continue
-                if not isinstance(chunk, bytes):
-                    chunk = chunk.encode("utf8")
-                len_str = hex(len(chunk))[2:]
-                to_send = bytearray(len_str.encode())
-                to_send += b"\r\n"
-                to_send += chunk
-                to_send += b"\r\n"
-                self.send(to_send)
-
-        # After the if clause, to always have a closed body
-        self.send(b"0\r\n\r\n")
+        self.request(method, url, body=body, headers=headers, chunked=True)
 
 
 class HTTPSConnection(HTTPConnection):
@@ -526,9 +544,9 @@ class HTTPSConnection(HTTPConnection):
             ca_cert_data=self.ca_cert_data,
             server_hostname=hostname,
             ssl_context=ssl_context,
+            assert_hostname=proxy_config.assert_hostname,
+            assert_fingerprint=proxy_config.assert_fingerprint,
             # Features that aren't implemented for proxies yet:
-            assert_fingerprint=None,
-            assert_hostname=None,
             cert_file=None,
             key_file=None,
             key_password=None,
@@ -608,13 +626,15 @@ def _ssl_wrap_socket_and_match_hostname(
     ):
         context.load_default_certs()
 
-    # Our upstream implementation of ssl.match_hostname()
-    # only applies this normalization to IP addresses so it doesn't
-    # match DNS SANs so we do the same thing!
+    # Ensure that IPv6 addresses are in the proper format and don't have a
+    # scope ID. Python's SSL module fails to recognize scoped IPv6 addresses
+    # and interprets them as DNS hostnames.
     if server_hostname is not None:
-        stripped_hostname = server_hostname.strip("[]")
-        if is_ipaddress(stripped_hostname):
-            server_hostname = stripped_hostname
+        normalized = server_hostname.strip("[]")
+        if "%" in normalized:
+            normalized = normalized[: normalized.rfind("%")]
+        if is_ipaddress(normalized):
+            server_hostname = normalized
 
     ssl_sock = ssl_wrap_socket(
         sock=sock,