+++ test/with_dummyserver/test_socketlevel.py
@@ -8,6 +8,7 @@ import select
 import shutil
 import socket
 import ssl
+import sys
 import tempfile
 import time
 from collections import OrderedDict
@@ -21,7 +22,7 @@ from test import (
     resolvesLocalhostFQDN,
 )
 from threading import Event
-from typing import Any, Generator, List, Optional
+from typing import Any, Callable, Generator, List, Optional
 from typing import OrderedDict as OrderedDictType
 from typing import Tuple, Union
 from unittest import mock
@@ -352,21 +353,21 @@ class TestClientCerts(SocketDummyServerTestCase):
     def test_load_keyfile_with_invalid_password(self) -> None:
         assert ssl_.SSLContext is not None
         context = ssl_.SSLContext(ssl_.PROTOCOL_SSLv23)
-
-        # Different error is raised depending on context.
-        if ssl_.IS_PYOPENSSL:
-            from OpenSSL.SSL import Error  # type: ignore[import]
-
-            expected_error = Error
-        else:
-            expected_error = ssl.SSLError
-
-        with pytest.raises(expected_error):
+        with pytest.raises(ssl.SSLError):
             context.load_cert_chain(
                 certfile=self.cert_path,
                 keyfile=self.password_key_path,
                 password=b"letmei",
             )
+
+    # For SecureTransport, the validation that would raise an error in
+    # this case is deferred.
+    @notSecureTransport()
+    def test_load_invalid_cert_file(self) -> None:
+        assert ssl_.SSLContext is not None
+        context = ssl_.SSLContext(ssl_.PROTOCOL_SSLv23)
+        with pytest.raises(ssl.SSLError):
+            context.load_cert_chain(certfile=self.password_key_path)
 
 
 class TestSocketClosing(SocketDummyServerTestCase):
@@ -1491,6 +1492,63 @@ class TestSSL(SocketDummyServerTestCase):
                 pool.request("GET", "/", retries=False, timeout=LONG_TIMEOUT)
         assert server_closed.wait(LONG_TIMEOUT), "The socket was not terminated"
 
+    # SecureTransport can read only small pieces of data at the moment.
+    # https://github.com/urllib3/urllib3/pull/2674
+    @notSecureTransport()
+    @pytest.mark.skipif(
+        os.environ.get("CI") == "true" and sys.implementation.name == "pypy",
+        reason="too slow to run in CI",
+    )
+    @pytest.mark.parametrize(
+        "preload_content,read_amt", [(True, None), (False, None), (False, 2**31)]
+    )
+    def test_requesting_large_resources_via_ssl(
+        self, preload_content: bool, read_amt: Optional[int]
+    ) -> None:
+        """
+        Ensure that it is possible to read 2 GiB or more via an SSL
+        socket.
+        https://github.com/urllib3/urllib3/issues/2513
+        """
+        content_length = 2**31  # (`int` max value in C) + 1.
+        ssl_ready = Event()
+
+        def socket_handler(listener: socket.socket) -> None:
+            sock = listener.accept()[0]
+            ssl_sock = ssl.wrap_socket(
+                sock,
+                server_side=True,
+                keyfile=DEFAULT_CERTS["keyfile"],
+                certfile=DEFAULT_CERTS["certfile"],
+                ca_certs=DEFAULT_CA,
+            )
+            ssl_ready.set()
+
+            while not ssl_sock.recv(65536).endswith(b"\r\n\r\n"):
+                continue
+
+            ssl_sock.send(
+                b"HTTP/1.1 200 OK\r\n"
+                b"Content-Type: text/plain\r\n"
+                b"Content-Length: %d\r\n\r\n" % content_length
+            )
+
+            chunks = 2
+            for i in range(chunks):
+                ssl_sock.sendall(bytes(content_length // chunks))
+
+            ssl_sock.close()
+            sock.close()
+
+        self._start_server(socket_handler)
+        ssl_ready.wait(5)
+        with HTTPSConnectionPool(
+            self.host, self.port, ca_certs=DEFAULT_CA, retries=False
+        ) as pool:
+            response = pool.request("GET", "/", preload_content=preload_content)
+            data = response.data if preload_content else response.read(read_amt)
+            assert len(data) == content_length
+
 
 class TestErrorWrapping(SocketDummyServerTestCase):
     def test_bad_statusline(self) -> None:
@@ -1642,6 +1700,75 @@ class TestHeaders(SocketDummyServerTestCase):
                 (k, v) for (k, v) in r.headers.items() if k.startswith("X-Header-")
             ]
             assert expected_response_headers == actual_response_headers
+
+    @pytest.mark.parametrize(
+        "method_type, body_type",
+        [
+            ("GET", None),
+            ("POST", None),
+            ("POST", "bytes"),
+            ("POST", "bytes-io"),
+        ],
+    )
+    def test_headers_sent_with_add(
+        self, method_type: str, body_type: Optional[str]
+    ) -> None:
+        """
+        Confirm that when adding headers with combine=True that we simply append to the
+        most recent value, rather than create a new header line.
+        """
+        body: Union[None, bytes, io.BytesIO]
+        if body_type is None:
+            body = None
+        elif body_type == "bytes":
+            body = b"my-body"
+        elif body_type == "bytes-io":
+            body = io.BytesIO(b"bytes-io-body")
+            body.seek(0, 0)
+        else:
+            raise ValueError("Unknonw body type")
+
+        buffer: bytes = b""
+
+        def socket_handler(listener: socket.socket) -> None:
+            nonlocal buffer
+            sock = listener.accept()[0]
+            sock.settimeout(0)
+
+            start = time.time()
+            while time.time() - start < (LONG_TIMEOUT / 2):
+                try:
+                    buffer += sock.recv(65536)
+                except OSError:
+                    continue
+
+            sock.sendall(
+                b"HTTP/1.1 200 OK\r\n"
+                b"Server: example.com\r\n"
+                b"Content-Length: 0\r\n\r\n"
+            )
+            sock.close()
+
+        self._start_server(socket_handler)
+
+        headers = HTTPHeaderDict()
+        headers.add("A", "1")
+        headers.add("C", "3")
+        headers.add("B", "2")
+        headers.add("B", "3")
+        headers.add("A", "4", combine=False)
+        headers.add("C", "5", combine=True)
+        headers.add("C", "6")
+
+        with HTTPConnectionPool(self.host, self.port, retries=False) as pool:
+            r = pool.request(
+                method_type,
+                "/",
+                body=body,
+                headers=headers,
+            )
+            assert r.status == 200
+            assert b"A: 1\r\nA: 4\r\nC: 3, 5\r\nC: 6\r\nB: 2\r\nB: 3" in buffer
 
 
 class TestBrokenHeaders(SocketDummyServerTestCase):
@@ -2035,12 +2162,12 @@ class TestContentFraming(SocketDummyServerTestCase):
         assert b"Transfer-Encoding: chunked\r\n" in sent_bytes
         assert b"User-Agent: python-urllib3/" in sent_bytes
         assert b"content-length" not in sent_bytes.lower()
-
-        # TODO: Remove the .lower() after solving #2515
-        assert b"\r\n\r\na\r\nxxxxxxxxxx\r\n0\r\n\r\n" in sent_bytes.lower()
+        assert b"\r\n\r\na\r\nxxxxxxxxxx\r\n0\r\n\r\n" in sent_bytes
 
     @pytest.mark.parametrize("method", ["POST", "PUT", "PATCH"])
-    @pytest.mark.parametrize("body_type", ["file", "generator", "bytes"])
+    @pytest.mark.parametrize(
+        "body_type", ["file", "generator", "bytes", "bytearray", "file_text"]
+    )
     def test_chunked_not_specified(self, method: str, body_type: str) -> None:
         buffer = bytearray()
 
@@ -2079,6 +2206,15 @@ class TestContentFraming(SocketDummyServerTestCase):
             body.seek(0, 0)
             should_be_chunked = True
 
+        elif body_type == "file_text":
+            body = io.StringIO("x" * 10)
+            body.seek(0, 0)
+            should_be_chunked = True
+
+        elif body_type == "bytearray":
+            body = bytearray(b"x" * 10)
+            should_be_chunked = False
+
         else:
             body = b"x" * 10
             should_be_chunked = False
@@ -2098,10 +2234,68 @@ class TestContentFraming(SocketDummyServerTestCase):
         if should_be_chunked:
             assert b"content-length" not in sent_bytes.lower()
             assert b"Transfer-Encoding: chunked\r\n" in sent_bytes
-            # TODO: Remove the .lower() after solving #2515
-            assert b"\r\n\r\na\r\nxxxxxxxxxx\r\n0\r\n\r\n" in sent_bytes.lower()
+            assert b"\r\n\r\na\r\nxxxxxxxxxx\r\n0\r\n\r\n" in sent_bytes
 
         else:
             assert b"Content-Length: 10\r\n" in sent_bytes
             assert b"transfer-encoding" not in sent_bytes.lower()
             assert sent_bytes.endswith(b"\r\n\r\nxxxxxxxxxx")
+
+    @pytest.mark.parametrize(
+        "header_transform",
+        [str.lower, str.title, str.upper],
+    )
+    @pytest.mark.parametrize(
+        ["header", "header_value", "expected"],
+        [
+            ("content-length", "10", b": 10\r\n\r\nxxxxxxxx"),
+            (
+                "transfer-encoding",
+                "chunked",
+                b": chunked\r\n\r\n8\r\nxxxxxxxx\r\n0\r\n\r\n",
+            ),
+        ],
+    )
+    def test_framing_set_via_headers(
+        self,
+        header_transform: Callable[[str], str],
+        header: str,
+        header_value: str,
+        expected: bytes,
+    ) -> None:
+        buffer = bytearray()
+
+        def socket_handler(listener: socket.socket) -> None:
+            nonlocal buffer
+            sock = listener.accept()[0]
+            sock.settimeout(0)
+
+            start = time.time()
+            while time.time() - start < (LONG_TIMEOUT / 2):
+                try:
+                    buffer += sock.recv(65536)
+                except OSError:
+                    continue
+
+            sock.sendall(
+                b"HTTP/1.1 200 OK\r\n"
+                b"Server: example.com\r\n"
+                b"Content-Length: 0\r\n\r\n"
+            )
+            sock.close()
+
+        self._start_server(socket_handler)
+
+        with HTTPConnectionPool(
+            self.host, self.port, timeout=LONG_TIMEOUT, retries=False
+        ) as pool:
+            resp = pool.request(
+                "POST",
+                "/",
+                body=b"xxxxxxxx",
+                headers={header_transform(header): header_value},
+            )
+            assert resp.status == 200
+
+            sent_bytes = bytes(buffer)
+            assert sent_bytes.endswith(expected)